clean renaming etc
