+++ packages/magnitude-core/src/common/actions.ts
@@ -1,5 +1,5 @@
 import { ClickWebAction, NavigateWebAction, ScrollWebAction, TypeWebAction } from "@/web/types";
-import { ClickIngredient, TypeIngredient } from "@/recipe/types";
+import { ClickIntent, TypeIntent } from "@/recipe/types";
 import { ScrollIngredient } from "@/ai/baml_client";
 
 export type ActionVariant = 'load' | 'click' | 'hover' | 'type' | 'scroll' | 'wait' | 'back';
@@ -7,8 +7,8 @@ export type ActionVariant = 'load' | 'click' | 'hover' | 'type' | 'scroll' | 'wa
 export type ActionDescriptor = NavigateActionDescriptor | ClickActionDescriptor | TypeActionDescriptor | ScrollActionDescriptor;
 
 export type NavigateActionDescriptor = NavigateWebAction & { screenshot: string };
-export type ClickActionDescriptor = ClickIngredient & ClickWebAction & { screenshot: string };
-export type TypeActionDescriptor = TypeIngredient & TypeWebAction & { screenshot: string };
+export type ClickActionDescriptor = ClickIntent & ClickWebAction & { screenshot: string };
+export type TypeActionDescriptor = TypeIntent & TypeWebAction & { screenshot: string };
 export type ScrollActionDescriptor = ScrollIngredient & ScrollWebAction & { screenshot: string };
 
 //const a: ClickActionDescriptor = { variant: 'click', x: 1, y: 2, target: ''}
