+++ packages/magnitude-core/src/common/index.ts
@@ -1,5 +1,4 @@
 export * from './actions';
 export * from './events';
 export * from './failure';
-export * from './state';
 export * from './util';
